// src/event-validation/registry.generated.ts
export const KIND_REGISTRY_SOURCE = "https://github.com/nostr-protocol/registry-of-kinds"
export const KIND_REGISTRY_COMMIT = "d51285770b3093e575bdf8d0bd81786fe1efbd6d"
export const KIND_REGISTRY_SHA256 = "4e7afedd1937c5728200da39906ed7e0c4f91409ad0a58a414fcd2eb356f6cac"
export const PINNED_KIND_REGISTRY: KindRegistrySchema = {
// ...
  "kinds": {
    "0": {
      "description": "User metadata",
      "in_use": true,
      "content": {
        "type": "json"
      },
      "multiple": [
        "emoji"
      ],
      "tags": [
        {
          "name": "emoji",
          "next": {
            "type": "free",
            "required": true,
            "next": {
              "type": "url",
              "required": true
            }
          },
          "description": "emojis references used inside the .content"
        }
      ]
    },
// ...
  }
}